orchids-challenge/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Website cloning API with scraping and AI capabilities"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "playwright>=1.42.0",
    "Pillow>=10.2.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
]
